+++ packages/paperdave-events/src/index.ts
@@ -1,4 +1,4 @@
-import type { Dict, EmptyObject } from '@davecode/types';
+import type { Dict, EmptyObject } from '@paperdave/types';
 
 export type BaseEventMap = Record<string, any> | Dict<unknown>;
 export type EventHandler<T> = undefined extends T ? (event: never) => void : (event: T) => void;